lab-interpreter/requirements.txt
streamlit==1.34.0
anthropic==0.25.0
PyMuPDF==1.24.1
fpdf2==2.7.0
pandas==2.1.4
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
